+++ projectionizer/tests/test_step_0_sample.py
@@ -15,9 +15,8 @@ def test_SampleChunk():
     with setup_tempdir('test_step_0') as tmp_folder:
         mock_path = os.path.join(tmp_folder, 'full-sample.feather')
         write_feather(mock_path, pd.DataFrame(np.arange(100), columns=['foo', ]))
-        params = {'circuit_config': 'circuit',
+        params = {'circuit_config': os.path.join(tmp_folder, 'CircuitConfig'),
                   'physiology_path': 'fake_string',
-                  'morphology_path': 'fake_string',
                   'folder': tmp_folder,
                   'sgid_offset': 0,
                   'oversampling': 0,